# Frontend/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "Gitlab Agent",
        "version": "0.1.0"
    },
    "paths": {
...
        "/api/v1/config/available-avatars": {
            "get": {
                "tags": ["config"],
                "summary": "Get Available Avatars",
                "description": "Get a list of available bot avatars.",
                "operationId": "get_available_avatars_api_v1_config_available_avatars_get",
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "additionalProperties": {
                                        "type": "string"
                                    },
                                    "type": "object",
                                    "title": "Response Get Available Avatars Api V1 Config Available Avatars Get"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponseModel"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Invalid credentials",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponseModel"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponseModel"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Resource not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponseModel"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponseModel"
                                }
                            }
                        }
                    },
                    "502": {
                        "description": "Bad Gateway",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponseModel"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/config/available-llms": {
            "get": {
                "tags": ["config"],
                "summary": "Get Available Llms",
                "description": "Get a list of available LLM models.",
                "operationId": "get_available_llms_api_v1_config_available_llms_get",
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "additionalProperties": {
                                        "$ref": "#/components/schemas/LLMModelInfo"
                                    },
                                    "type": "object",
                                    "title": "Response Get Available Llms Api V1 Config Available Llms Get"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponseModel"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Invalid credentials",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponseModel"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponseModel"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Resource not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponseModel"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponseModel"
                                }
                            }
                        }
                    },
                    "502": {
                        "description": "Bad Gateway",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponseModel"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/config/": {
            "get": {
                "tags": ["config"],
                "summary": "Get Configs",
                "description": "Get the current configuration settings.",
                "operationId": "get_configs_api_v1_config__get",
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Configs"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponseModel"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Invalid credentials",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponseModel"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponseModel"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Resource not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponseModel"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponseModel"
                                }
                            }
                        }
                    },
                    "502": {
                        "description": "Bad Gateway",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponseModel"
                                }
                            }
                        }
                    }
                }
            },
            "patch": {
                "tags": ["config"],
                "summary": "Update Configs",
                "description": "Update the configuration settings.",
                "operationId": "update_configs_api_v1_config__patch",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ConfigsUpdate"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Configs"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponseModel"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Invalid credentials",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponseModel"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponseModel"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Resource not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponseModel"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponseModel"
                                }
                            }
                        }
                    },
                    "502": {
                        "description": "Bad Gateway",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponseModel"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "HTTPBearer": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Configs": {
                "properties": {
                    "max_chat_history": {
                        "type": "integer",
                        "title": "Max Chat History"
                    },
                    "max_tokens_per_diff": {
                        "type": "integer",
                        "title": "Max Tokens Per Diff"
                    },
                    "max_tokens_per_context": {
                        "type": "integer",
                        "title": "Max Tokens Per Context"
                    },
                    "default_llm_model": {
                        "type": "string",
                        "title": "Default Llm Model"
                    },
                    "avatar_default_name": {
                        "type": "string",
                        "title": "Avatar Default Name"
                    },
                    "available_llms": {
                        "additionalProperties": {
                            "$ref": "#/components/schemas/LLMModelInfo"
                        },
                        "type": "object",
                        "title": "Available Llms"
                    }
                },
                "type": "object",
                "required": [
                    "max_chat_history",
                    "max_tokens_per_diff",
                    "max_tokens_per_context",
                    "default_llm_model",
                    "avatar_default_name"
                ],
                "title": "Configs"
            },
            "ConfigsUpdate": {
                "properties": {
                    "max_chat_history": {
                        "anyOf": [
                            {
                                "type": "integer"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Max Chat History"
                    },
                    "max_tokens_per_diff": {
                        "anyOf": [
                            {
                                "type": "integer"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Max Tokens Per Diff"
                    },
                    "max_tokens_per_context": {
                        "anyOf": [
                            {
                                "type": "integer"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Max Tokens Per Context"
                    },
                    "default_llm_model": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Default Llm Model"
                    },
                    "avatar_default_name": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Avatar Default Name"
                    }
                },
                "type": "object",
                "title": "ConfigsUpdate"
            },
            "ErrorResponseModel": {
                "properties": {
                    "detail": {
                        "type": "string",
                        "title": "Detail"
                    }
                },
                "type": "object",
                "required": ["detail"],
                "title": "ErrorResponseModel",
                "description": "Standard error response model."
            },
...
            "HTTPValidationError": {
                "properties": {
                    "detail": {
                        "items": {
                            "$ref": "#/components/schemas/ValidationError"
                        },
                        "type": "array",
                        "title": "Detail"
                    }
                },
                "type": "object",
                "title": "HTTPValidationError"
            },
            "LLMModelInfo": {
                "properties": {
                    "model_name": {
                        "type": "string",
                        "title": "Model Name"
                    },
                    "context_window": {
                        "type": "integer",
                        "title": "Context Window"
                    },
                    "max_output_tokens": {
                        "type": "integer",
                        "title": "Max Output Tokens"
                    },
                    "temperature": {
                        "type": "number",
                        "title": "Temperature"
                    },
                    "top_p": {
                        "type": "number",
                        "title": "Top P",
                        "default": 0.95
                    },
                    "input_token_cost": {
                        "type": "number",
                        "title": "Input Token Cost",
                        "default": 0.0
                    },
                    "output_token_cost": {
                        "type": "number",
                        "title": "Output Token Cost",
                        "default": 0.0
                    },
                    "additional_kwargs_schema": {
                        "additionalProperties": true,
                        "type": "object",
                        "title": "Additional Kwargs Schema"
                    }
                },
                "type": "object",
                "required": [
                    "model_name",
                    "context_window",
                    "max_output_tokens",
                    "temperature"
                ],
                "title": "LLMModelInfo"
            },
...
            "ValidationError": {
                "properties": {
                    "loc": {
                        "items": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "integer"
                                }
                            ]
                        },
                        "type": "array",
                        "title": "Location"
                    },
                    "msg": {
                        "type": "string",
                        "title": "Message"
                    },
                    "type": {
                        "type": "string",
                        "title": "Error Type"
                    }
                },
                "type": "object",
                "required": ["loc", "msg", "type"],
                "title": "ValidationError"
            },
...
        },
        "securitySchemes": {
            "HTTPBearer": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    }
}